+++ src/main/java/com/example/demo/service/Validationservice.java
@@ -3,7 +3,7 @@ import java.util.List;
 import com.example.demo.entity.ValidationEntity;
 public interface Validationservice{
 ValidationEntity postdata(ValidationEntity entity);
-ValidationEntity getData(int id);
+ValidationEntity getData(Long id);
 
 }
